requirements.txt
amqp==2.3.2
astroid==2.0.4
beautifulsoup4==4.6.0
billiard==3.5.0.4
CacheControl==0.12.5
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
Django==1.11.4
django-celery-beat==1.1.1
django-crontab==0.7.1
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.8.2
firebase-admin==2.13.0
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.11.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.14.2
gunicorn==19.7.1
idna==2.7
isort==4.3.4
kombu==4.2.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
msgpack==0.5.6
protobuf==3.6.1
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==2.1.1
pytz==2018.5
requests==2.19.1
robobrowser==0.5.3
rsa==3.4.2
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
whitenoise==4.0
wrapt==1.10.11
